+++ src/main/scala/org/programmiersportgruppe/redis/replies.scala
@@ -7,6 +7,6 @@ sealed abstract class Reply
 
 case class StatusReply(status: String) extends Reply
 case class ErrorReply(error: String) extends Reply
-case class IntegerReply(value: Int) extends Reply
+case class IntegerReply(value: Long) extends Reply
 case class BulkReply(data: Option[ByteString]) extends Reply
 case class MultiBulkReply(replies: Seq[Reply]) extends Reply
